+++ packages/emp/src/globals.d.ts
@@ -2,3 +2,5 @@ declare module 'webpack-federated-stats-plugin'
 declare module 'esbuild-css-modules-plugin'
 declare module 'inquirer'
 declare module 'git-promise'
+
+declare module 'swc-plugin-transform-import'