+++ projects/antd-base/emp-config.js
@@ -3,7 +3,10 @@ const {defineConfig} = require('@efox/emp')
 module.exports = defineConfig({
   //   splitCss: false,
   // plugins: [pluginBabelReact],
-  moduleTransform: {parser: 'swc'},
+  moduleTransform: {
+    parser: 'swc',
+    // antdTransformImport: false,
+  },
   build: {
     // target: 'es2021',
   },