+++ projects/antd-base/src/bootstrap.tsx
@@ -1,5 +1,5 @@
 import {render} from 'react-dom'
-// import Layout from './AntdLayout'
-import Layout from './Layout'
+import Layout from './AntdLayout'
+// import Layout from './Layout'
 
 render(<Layout />, document.getElementById('emp-root'))